Create README.md
Just decided to join the human race. Am trying to learn how to do videos and editing. My wish is to educate and make people laugh because the world is too serious right now and people need to be able to escape the madness if but for a brief moment or two.